+++ source/extensions/omni.isaac.extension_templates/omni/isaac/extension_templates/extension.py
@@ -18,20 +18,9 @@ import omni.kit.commands
 import omni.timeline
 import omni.ui as ui
 import omni.usd
-from omni.isaac.ui.element_wrappers import CollapsableFrame, TextBlock
+from omni.isaac.ui.element_wrappers import CollapsableFrame, ScrollingWindow, TextBlock
 from omni.isaac.ui.menu import make_menu_item_description
-from omni.isaac.ui.ui_utils import (
-    add_line_rect_flourish,
-    btn_builder,
-    color_picker_builder,
-    float_builder,
-    get_style,
-    int_builder,
-    setup_ui_headers,
-    state_btn_builder,
-    str_builder,
-    xyz_builder,
-)
+from omni.isaac.ui.ui_utils import btn_builder, get_style, setup_ui_headers, str_builder
 from omni.kit.menu.utils import add_menu_items, remove_menu_items
 
 from .template_generator import TemplateGenerator
@@ -49,7 +38,7 @@ class Extension(omni.ext.IExt):
         self._usd_context = omni.usd.get_context()
 
         # Build Window
-        self._window = ui.Window(
+        self._window = ScrollingWindow(
             title=EXTENSION_NAME, width=600, height=500, visible=False, dockPreference=ui.DockPreference.LEFT_BOTTOM
         )
         self._window.set_visibility_changed_fn(self._on_window)